(validator, miner, executor), DAH-1651, add CI tests.
