+++ neurons/miners/pyproject.toml
@@ -24,6 +24,7 @@ dependencies = [
     "tenacity>=9.1.2",
     "uvicorn>=0.35.0",
     "pytest>=8.4.2",
+    "httpx>=0.28.0",
 ]
 requires-python = "<3.13,>=3.11"
 readme = "README.md"
@@ -36,3 +37,7 @@ build-backend = "pdm.backend"
 
 [tool.pdm]
 distribution = true
+
+[tool.pytest.ini_options]
+pythonpath = ["src"]
+testpaths = ["tests"]
