+++ neurons/validators/pyproject.toml
@@ -49,7 +49,7 @@ build-backend = "pdm.backend"
 distribution = true
 
 [tool.pytest.ini_options]
-pythonpath = ["src"]
+pythonpath = ["src", "."]
 testpaths = ["tests"]
 
 [dependency-groups]
